started implementation for scenario and test with JSON files. Minor refactoring in the code.

+++ src/main/java/fr/elpresidente/game/factions/FactionController.java
@@ -62,7 +62,7 @@ public class FactionController {
 
     }
 
-    public double determineTotalSupporters() {
+    public int determineTotalSupporters() {
         return this.factions.getFactionList()
                 .stream().mapToInt(faction -> faction.getSupporters()).sum();
     }
